Revised the InitializeDb function signature
When running tests, for non existent records the logs would be
printed:

2025/02/28 09:42:48 /Users/alex/dev/go/repos/planetgolang.dev/database.go:41 record not found
[0.054ms] [rows:0] SELECT * FROM `posts` WHERE url = "https://example.com/go-gin-rest-api" ORDER BY `posts`.`id` LIMIT 1

In order to silence them, I've revised the signature of the
InitializeDb, so that it can explicitly take the database name and the
configuration which is to be passed to gorm, when opening the
database.

My reasoning is that it give us flexibility, so that we can specify
what file name we need, for example in test scenarios. As well as
being able to modify gorm options, like in the above case, where we
don't want logs to be printed out during testing. Since they make the
test output look noisy & messy.